Update GetUsageSummary example to show AdditionalProperties access at all layers
